convert records table
